+++ src/store/selector/selector.js
@@ -57,8 +57,11 @@ const selectedStatus = (arrayStatus) => {
 
 const filteredNameOrOrder = (nameOrOrder) => {
   return (order, name) => {
+    // if(Number.isFinite(Number(nameOrOrder))){
+    //   return order
+    // }
     const getValue = Number.isFinite(Number(nameOrOrder))
-      ? order.includes(nameOrOrder)
+      ? order.startsWith(nameOrOrder)
       : name.toLowerCase().includes(nameOrOrder);
     return getValue;
   };
